Feat: 링크 수정, 삭제 기능 추가. 태그 기능과 Tag 페이지 추가. All 페이지에 검색 기능 추가 - 링크 수정   - KEEP 시 이미 저장된 링크라면 수정할 수 있도록 기능 추가
  - linkTicket 클릭 시 해당 링크 수정할 수 있도록 기능 추가
  - linkTicket 마우스 오버 시 style 추가

  - 링크 저장, 수정 시 태그도 함께 저장 가능하며 Dropdown에 유저의 태그 리스트 출력
  - 링크 수정 모달에 삭제 버튼 추가

- Tag 페이지 추가
  - AllLayout으로 All과 Tag 페이지의 공통 레이아웃 분리
  - 태그 검색 기능 추가
  - 태그 출력 시 읽은 태그, 안읽은 태그 분류

- All 페이지에 검색 기능 추가

+++ src/components/Form/TextInput/index.tsx
@@ -15,30 +15,44 @@ import ICON_ALL_DELETE from '@/assets/images/ic-all-delete.svg';
 //    input : 흰색
 //    button : 글자색=검정, 배경색=회색, 폰트=fontR
 
+// bgGray : input 컬러를 basicBg로 적용합니다.
 // point-button : submit 버튼의 폰트를 pointFont로 적용합니다.
 
 type Props = {
-  className?: 'point-button';
+  className?: 'point-button' | 'bgGray' | 'point-button bgGray';
   placeholderText: string;
   buttonText: string;
   onClick: (value: string) => void;
+  onClickDeleteAll?: () => void;
   style?: { [attribute: string]: string };
   onChange?: (value: string) => string; // 값에 변화를 주고싶으면 onChange를 전달합니다
+  deleteAfterSubmit?: boolean; // default: true. 제출후 input을 비우고 싶지 않다면 false 주세요
 };
 
-export default function TextForm({ className, placeholderText, buttonText, onClick, onChange, style }: Props) {
+export default function TextInput({
+  className,
+  placeholderText,
+  buttonText,
+  onClick,
+  onClickDeleteAll,
+  onChange,
+  style,
+  deleteAfterSubmit = true,
+}: Props) {
   const [value, setValue] = useState('');
 
   const onClickForSubmit = () => {
     onClick(value);
-    onDeleteAll();
+    if (deleteAfterSubmit) {
+      onDeleteAll();
+    }
   };
 
   const onDeleteAll = () => {
     setValue('');
   };
   return (
-    <S.Form className={className + (value !== '' ? 'isValue' : '')} style={style}>
+    <S.TextInput className={className + (value !== '' ? ' isValue' : '')} style={style}>
       <div className="input-wrapper">
         <input
           type="text"
@@ -63,6 +77,9 @@ export default function TextForm({ className, placeholderText, buttonText, onCli
           type="button"
           onClick={() => {
             onDeleteAll();
+            if (onClickDeleteAll) {
+              onClickDeleteAll();
+            }
           }}
         >
           <img src={ICON_ALL_DELETE} alt="delete-icon" height={22} width={22} />
@@ -78,6 +95,6 @@ export default function TextForm({ className, placeholderText, buttonText, onCli
       >
         {buttonText}
       </button>
-    </S.Form>
+    </S.TextInput>
   );
 }
